selenium-tests/requirements.txt
# Selenium WebDriver and related tools
selenium>=4.15.0
webdriver-manager>=4.0.0

# Testing framework
pytest>=7.4.0
pytest-html>=3.2.0
pytest-xdist>=3.3.0
pytest-json-report>=1.5.0
pytest-cov>=4.1.0

# Allure reporting (optional)
allure-pytest>=2.13.0

# Data generation and manipulation
faker>=20.1.0
python-dotenv>=1.0.0

# Database connectivity (MongoDB)
pymongo>=4.5.0

# HTTP requests for API testing
requests>=2.31.0

# Data validation and parsing
pydantic>=2.5.0

# Excel/CSV data handling
openpyxl>=3.1.0
pandas>=2.1.0

# Image processing for screenshot comparison
Pillow>=10.1.0

# Configuration management
pyyaml>=6.0.1

# Utilities
colorama>=0.4.6
rich>=13.7.0
click>=8.1.0

# Development tools
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0

# Additional utilities
parameterized>=0.9.0
retry>=0.9.2

# Performance monitoring
psutil>=5.9.0

# Network utilities
netifaces>=0.11.0 